requirements.txt
torch
transformers
chromadb
Pillow
gradio
umap-learn
plotly
